requirements.txt
click
Pillow
pyheif
opencv-python
tqdm
